Moved ArrowDecorator into layers/utils
